+++ ingestion/src/main/java/feast/ingestion/transform/specs/FilterRelevantFunction.java
@@ -16,7 +16,7 @@
  */
 package feast.ingestion.transform.specs;
 
-import feast.common.function.Store;
+import feast.common.models.Store;
 import feast.proto.core.FeatureSetProto;
 import feast.proto.core.SourceProto;
 import feast.proto.core.StoreProto;
