+++ ingestion/src/main/java/feast/ingestion/values/FeatureSet.java
@@ -16,7 +16,7 @@
  */
 package feast.ingestion.values;
 
-import static feast.common.function.FeatureSet.getFeatureSetStringRef;
+import static feast.common.models.FeatureSet.getFeatureSetStringRef;
 import static feast.ingestion.utils.SpecUtil.getFieldsByName;
 
 import feast.proto.core.FeatureSetProto;
